Changed router file to display create profile form
